cleanup

+++ main.py
@@ -2,19 +2,18 @@ from flask import Flask, request, render_template
 from io import BytesIO
 from PyPDF2 import PdfReader
 import pandas as pd
-from openai.embeddings_utils import get_embedding, cosine_similarity
-import openai
 import os
 import requests
 from flask_cors import CORS
 from _md5 import md5
 from google.cloud import storage
+import base64
 
 app = Flask(__name__)
-# db=redis.from_url(os.environ['REDISCLOUD_URL'])
-# db = redis.StrictRedis(host='localhost', port=6379, db=0)
-# os.environ['CLOUD_STORAGE_BUCKET'] = 'researchgpt.appspot.com'
+
+os.environ['CLOUD_STORAGE_BUCKET'] = 'researchgpt.appspot.com'
 CLOUD_STORAGE_BUCKET = os.environ['CLOUD_STORAGE_BUCKET']
+
 CORS(app)
 
 class Chatbot():
@@ -80,109 +79,38 @@ class Chatbot():
                 continue
             filtered_pdf.append(row)
         df = pd.DataFrame(filtered_pdf)
-        print(df.columns)
         # remove elements with identical df[text] and df[page] values
-        # df = df.drop_duplicates(subset=['text', 'page'], keep='first')
+        df = df.drop_duplicates(subset=['text', 'page'], keep='first')
         # df['length'] = df['text'].apply(lambda x: len(x))
         print('Done creating dataframe')
         return df
-
-    def embeddings(self, df):
-        print('Calculating embeddings')
-        openai.api_key = os.getenv('OPENAI_API_KEY')
-        embedding_model = "text-embedding-ada-002"
-        embeddings = df.text.apply([lambda x: get_embedding(x, engine=embedding_model)])
-        df["embeddings"] = embeddings
-        print('Done calculating embeddings')
-        return df
-
-    def search(self, df, query, n=3, pprint=True):
-        query_embedding = get_embedding(
-            query,
-            engine="text-embedding-ada-002"
-        )
-        df["similarity"] = df.embeddings.apply(lambda x: cosine_similarity(x, query_embedding))
-        
-        results = df.sort_values("similarity", ascending=False, ignore_index=True)
-        # make a dictionary of the the first three results with the page number as the key and the text as the value. The page number is a column in the dataframe.
-        results = results.head(n)
-        global sources 
-        sources = []
-        for i in range(n):
-            # append the page number and the text as a dict to the sources list
-            sources.append({'Page '+str(results.iloc[i]['page']): results.iloc[i]['text'][:150]+'...'})
-        print(sources)
-        return results.head(n)
-    
-    def create_prompt(self, df, user_input):
-        result = self.search(df, user_input, n=3)
-        print(result)
-        prompt = """You are a large language model whose expertise is reading and summarizing scientific papers. 
-        You are given a query and a series of text embeddings from a paper in order of their cosine similarity to the query.
-        You must take the given embeddings and return a very detailed summary of the paper that answers the query.
-            
-            Given the question: """+ user_input + """
-            
-            and the following embeddings as data: 
-            
-            1.""" + str(result.iloc[0]['text']) + """
-            2.""" + str(result.iloc[1]['text']) + """
-            3.""" + str(result.iloc[2]['text']) + """
-
-            Return a detailed answer based on the paper:"""
-
-        print('Done creating prompt')
-        return prompt
-
-    def gpt(self, prompt):
-        print('Sending request to GPT-3')
-        openai.api_key = os.getenv('OPENAI_API_KEY')
-        r = openai.Completion.create(model="text-davinci-003", prompt=prompt, temperature=0.4, max_tokens=1500)
-        answer = r.choices[0]['text']
-        print('Done sending request to GPT-3')
-        response = {'answer': answer, 'sources': sources}
-        return response
 
 @app.route("/", methods=["GET", "POST"])
 def index():
     return render_template("index.html")
 
-
-@app.route("/process_pdf", methods=['POST'])
-def process_pdf():
-    print("Processing pdf")
-    print(request)
-    # print('the data')
-    # print(request.data)
-    file = request.data
-
-    key = md5(file).hexdigest()
+# a route to get the dataframe from the database
+@app.route("/get_df", methods=['POST'])
+def get_df():
+    print('Getting dataframe')
+    key = request.json['key']
     print(key)
+
+    query = request.json['query']
+    print(query)
+    
     # Create a Cloud Storage client.
     gcs = storage.Client()
     name = key+'.json'
-
-    # Get the bucket that the file will be uploaded to.
     bucket = gcs.get_bucket(CLOUD_STORAGE_BUCKET)
-    # Check if the file already exists
-    if bucket.blob(name).exists():
-        print("File already exists")
-        print("Done processing pdf")
-        return {"key": key}
-
-    pdf = PdfReader(BytesIO(file))
-    chatbot = Chatbot()
-    paper_text = chatbot.extract_text(pdf)
-    df = chatbot.create_df(paper_text)
-    df = chatbot.embeddings(df)
-    
-    # Create a new blob and upload the file's content.
     blob = bucket.blob(name)
-    blob.upload_from_string(df.to_json(), content_type='application/json')
-    # if db.get(key) is None:
-    #     db.set(key, df.to_json())
+    df = pd.read_json(BytesIO(blob.download_as_string()))
+
+    print('Done getting dataframe')
+    json_str = df.to_json(orient='records')
+    # print(len(json_str))
     print("Done processing pdf")
-    return {"key": key}
+    return {"df": json_str}
 
 # re-writing process_pdf to to just create the dataframe and send it to the frontend
 @app.route("/process_pdf", methods=['POST'])
@@ -196,25 +124,48 @@ def process_pdf():
     key = md5(file).hexdigest()
     print(key)
 
+    # if db.get(key) is not None:
+    #     print("File already exists")
+    #     return {"key": key, "exists": True}
+
+    # Create a Cloud Storage client.
+    gcs = storage.Client()
+    name = key+'.json'
+
+    # Get the bucket that the file will be uploaded to.
+    bucket = gcs.get_bucket(CLOUD_STORAGE_BUCKET)
+    # Check if the file already exists
+    if bucket.blob(name).exists():
+        print("File already exists")
+        print("Done processing pdf")
+        return {"key": key, "exists": True}
+
     pdf = PdfReader(BytesIO(file))
     chatbot = Chatbot()
     paper_text = chatbot.extract_text(pdf)
     df = chatbot.create_df(paper_text)
 
     json_str = df.to_json(orient='records')
-
+    # print(len(json_str))
     print("Done processing pdf")
     return {"key": key, "df": json_str}
 
 # a function save that takes in a dataframe and saves it to gcs
 @app.route("/save", methods=['POST'])
-def save_df_to_gcs():
+def save():
     print("Saving df to gcs")
-    print(request.json)
-    print(request.json['df'])
-    print(request.json['key'])
+    # print(request.json)
+    # print(request.json['df'])
+    # print(request.json['key'])
     df = request.json['df']
     key = request.json['key']
+
+    df = pd.DataFrame.from_dict(df)
+    
+    # Return error if the dataframe is empty
+    if df.empty:
+        return {"error": "No data found"}
+
     # Create a Cloud Storage client.
     gcs = storage.Client()
     name = key+'.json'
@@ -225,43 +176,45 @@ def save_df_to_gcs():
     if bucket.blob(name).exists():
         print("File already exists")
         print("Done processing pdf")
-        return {"key": key}
+        return {"key": key, "exists": True}
 
     # Create a new blob and upload the file's content.
     blob = bucket.blob(name)
     blob.upload_from_string(df.to_json(), content_type='application/json')
+
     # if db.get(key) is None:
     #     db.set(key, df.to_json())
-    print("Done processing pdf")
-    return {"key": key}
-
-# a function prompt that takes in a dataframe and a user input and returns a prompt
-@app.route("/prompt", methods=['POST'])
-def prompt():
-    print("Creating prompt")
-    print(request.json)
-    print(request.json['df'])
-    print(request.json['user_input'])
-    df = request.json['df']
-    user_input = request.json['user_input']
-    chatbot = Chatbot()
-    prompt = chatbot.create_prompt(df, user_input)
-    print("Done creating prompt")
-    return {"prompt": prompt}
+    # else:
+    #     print("File already exists")
+    #     return {"key": key, "exists": True}
+    
+    print("Saved pdf")
+    return {"key": key, "exists": False}
 
 @app.route("/download_pdf", methods=['POST'])
 def download_pdf():
     print("Downloading pdf")
-    print(request.json)
-    print(request.json['url'])
+    # print(request.json)
     chatbot = Chatbot()
     url = request.json['url']
-    r = requests.get(str(url))    
+    data = request.json['data']
+    data = base64.b64decode(data)
+
+    print(url)
+
+    # r = requests.get(str(url), allow_redirects=True, headers={
+    #     "Origin": "appspot.com", 
+    #     'Access-Control-Allow-Origin': '*',
+    #     'Access-Control-Allow-Methods': 'GET, POST, OPTIONS',
+    #     'Access-Control-Allow-Headers': 'Content-Type, Authorization'})    
+
     print("Downloading pdf")
-    print(r.status_code)
-    print(r.content)
-    # print(r.content)
-    key = md5(r.content).hexdigest()
+    key = md5(data).hexdigest()
+    print(key)
+
+    # if db.get(key) is not None:
+    #     print("File already exists")
+    #     return {"key": key, "exists": True}
 
     # Create a Cloud Storage client.
     gcs = storage.Client()
@@ -273,36 +226,16 @@ def download_pdf():
     if bucket.blob(name).exists():
         print("File already exists")
         print("Done processing pdf")
-        return {"key": key}
+        return {"key": key, "exists": True}
 
-    pdf = PdfReader(BytesIO(r.content))
+    pdf = PdfReader(BytesIO(data))
     paper_text = chatbot.extract_text(pdf)
     df = chatbot.create_df(paper_text)
-    df = chatbot.embeddings(df)
 
-    # Create a new blob and upload the file's content.
-    blob = bucket.blob(name)
-    blob.upload_from_string(df.to_json(), content_type='application/json')
+    json_str = df.to_json(orient='records')
+    # print(len(json_str))
     print("Done processing pdf")
-    return {"key": key}
-
-@app.route("/reply", methods=['POST'])
-def reply():
-    chatbot = Chatbot()
-    print(request.json)
-    query = request.json['query']
-    query = str(query)
-    print(query)
-    key = request.json['key']
-    gcs = storage.Client()
-    bucket = gcs.get_bucket(CLOUD_STORAGE_BUCKET)
-    blob = bucket.blob(key+'.json')
-    df = pd.read_json(BytesIO(blob.download_as_string()))
-    print(df.head(5))
-    prompt = chatbot.create_prompt(df, query)
-    response = chatbot.gpt(prompt)
-    print(response)
-    return response, 200
+    return {"key": key, "df": json_str}
 
 if __name__ == '__main__':
     app.run(host='127.0.0.1', port=8080, debug=True)
